+++ SessionMessagingKit/Calls/CallManagerProtocol.swift
@@ -9,6 +9,10 @@ public protocol CallManagerProtocol {
     func reportCurrentCallEnded(reason: CXCallEndedReason?)
     
     func showCallUIForCall(caller: String, uuid: String, mode: CallMode, interactionId: Int64?)
+    func handleICECandidates(message: CallMessage, sdpMLineIndexes: [UInt32], sdpMids: [String])
     func handleAnswerMessage(_ message: CallMessage)
+    
+    func currentWebRTCSessionMatches(callId: String) -> Bool
+    
     func dismissAllCallUI()
 }
