+++ SessionMessagingKit/Calls/CurrentCallProtocol.swift
@@ -7,7 +7,6 @@ import WebRTC
 public protocol CurrentCallProtocol {
     var uuid: String { get }
     var callId: UUID { get }
-    var webRTCSession: WebRTCSession { get }
     var hasStartedConnecting: Bool { get set }
     var hasEnded: Bool { get set }
