+++ SessionMessagingKit/Sending & Receiving/Message Handling/MessageReceiver+Calls.swift
@@ -23,19 +23,11 @@ extension MessageReceiver {
             case .provisionalAnswer: break // TODO: Implement
                 
             case let .iceCandidates(sdpMLineIndexes, sdpMids):
-                guard let currentWebRTCSession = WebRTCSession.current, currentWebRTCSession.uuid == message.uuid else {
-                    return
-                }
-                var candidates: [RTCIceCandidate] = []
-                let sdps = message.sdps
-                for i in 0..<sdps.count {
-                    let sdp = sdps[i]
-                    let sdpMLineIndex = sdpMLineIndexes[i]
-                    let sdpMid = sdpMids[i]
-                    let candidate = RTCIceCandidate(sdp: sdp, sdpMLineIndex: Int32(sdpMLineIndex), sdpMid: sdpMid)
-                    candidates.append(candidate)
-                }
-                currentWebRTCSession.handleICECandidates(candidates)
+                SessionEnvironment.shared?.callManager.wrappedValue?.handleICECandidates(
+                    message: message,
+                    sdpMLineIndexes: sdpMLineIndexes,
+                    sdpMids: sdpMids
+                )
                 
             case .endCall: MessageReceiver.handleEndCallMessage(db, message: message)
         }
@@ -150,9 +142,8 @@ extension MessageReceiver {
         SNLog("[Calls] Received answer message.")
         
         guard
-            let currentWebRTCSession: WebRTCSession = WebRTCSession.current,
-            currentWebRTCSession.uuid == message.uuid,
             let callManager: CallManagerProtocol = SessionEnvironment.shared?.callManager.wrappedValue,
+            callManager.currentWebRTCSessionMatches(callId: message.uuid),
             var currentCall: CurrentCallProtocol = callManager.currentCall,
             currentCall.uuid == message.uuid,
             let sender: String = message.sender
@@ -177,8 +168,8 @@ extension MessageReceiver {
         SNLog("[Calls] Received end call message.")
         
         guard
-            WebRTCSession.current?.uuid == message.uuid,
             let callManager: CallManagerProtocol = SessionEnvironment.shared?.callManager.wrappedValue,
+            callManager.currentWebRTCSessionMatches(callId: message.uuid),
             let currentCall: CurrentCallProtocol = callManager.currentCall,
             currentCall.uuid == message.uuid,
             let sender: String = message.sender
